Split decorators into files
