+++ tests/front/acceptance/cucumber/decorators/catalog-volume/report.js
@@ -1,0 +1,16 @@
+const Header = require('../common/header');
+
+const Report = async (nodeElement, createElementDecorator) => {
+    const children = {
+        'Header':  {
+            selector: '.AknTitleContainer',
+            decorator: Header
+        }
+    };
+    const getChildren = createElementDecorator(children, nodeElement);
+    const getHeader = async () => await getChildren('Header');
+
+    return { getHeader };
+};
+
+module.exports = Report;
